rename file to letter and hypens for consistency, https://github.com/fahreziadh/evaly/pull/24#issuecomment-2842969504
