+++ src/app/[locale]/(public)/pricing/page.tsx
@@ -1,3 +1,3 @@
-import ComingSoonPage from "../_components/ComingSoon";
+import ComingSoonPage from "../_components/coming-soon";
 
 export default ComingSoonPage;
