+++ src/app/[locale]/(public)/schools/page.tsx
@@ -1,3 +1,3 @@
-import ComingSoonPage from "../_components/ComingSoon";
+import ComingSoonPage from "../_components/coming-soon";
 
 export default ComingSoonPage;
